clean up users tests
