+++ templates/users/profile.html
@@ -13,11 +13,11 @@
 		</div>
 		<div class="lite-app-bar__controls">
             {% if super_user %}
-                <a href="{% url 'users:edit' data.user.id %}" type="submit" class="govuk-button">
+                <a href="{% url 'users:edit' data.user.id %}" type="submit" class="govuk-button" id="edit_button">
                     {% lcs 'USER_PROFILE_PAGE_EDIT' %}
                 </a>
             {% elif not data.user.role.id == super_user_role_id %}
-                <a href="{% url 'users:edit' data.user.id %}" type="submit" class="govuk-button">
+                <a href="{% url 'users:edit' data.user.id %}" type="submit" class="govuk-button" id="edit_button">
                     {% lcs 'USER_PROFILE_PAGE_EDIT' %}
                 </a>
             {% endif %}
