+++ ui_automation_tests/features/users.feature
@@ -1,8 +1,6 @@
 @all @users
 Feature: I want to test users
 
-  NOTE: I run the manage users test needs to get edited to just do edit user.
-
   @deactivate_user @regression
   Scenario: Add user, deactivate then reactivate
     Given I sign in to SSO or am signed into SSO
@@ -26,6 +24,10 @@ Feature: I want to test users
     When filter status has been changed to "All"
     Then I see new user
 
-  @manage_users @smoke
-  Scenario: Manage user
-    Given I run the manage users test
+  @edit_users @smoke
+  Scenario: Edit user
+    Given I sign in to SSO or am signed into SSO
+    And I go to users
+    When I add a new user
+    And I edit new user
+    Then I see new user
